backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
alembic==1.15.2
asyncpg==0.30.0
pydantic==2.11.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
boto3==1.37.12
cryptography==44.0.2
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
